+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ### ? - ?
 
+##### Additions :tada:
+
+- Tileset and raster overlay credits are now shown in Editor viewports.
 - Added point cloud shading options to `Cesium3DTileset`, which allow point cloud tilesets to be rendered with attenuation based on geometric error.
 
 ### v1.25.1 - 2023-05-02